aksh-async/requirements.txt
aiohttp[speedups]<3.8
aiofile<3.6
fastapi<0.64
pydantic<1.9
python-magic<0.5
textract<1.7
uvicorn[standard]<0.14
